Split rdf-components-mocha into separate files
